Refactor out to test/common
