+++ package.json
@@ -81,8 +81,8 @@
     "build:plugins": "npm run build:plugin:oni-plugin-typescript",
     "build:plugin:oni-plugin-typescript": "cd vim/core/oni-plugin-typescript && npm run build",
     "build:test": "cd test && tsc -p tsconfig.json",
-    "pack": "build --publish never",
     "copy-icons": "node build/CopyIcons.js",
+    "demo": "npm run build:test && mocha -t 30000 lib_test/test/Demo.js",
     "dist:win": "build --arch ia32 --publish never",
     "pack:win": "node build/BuildSetupTemplate.js && innosetup-compiler dist/setup.iss --verbose --O=dist",
     "test": "npm run test:unit && npm run test:integration",
@@ -98,6 +98,7 @@
     "lint:browser": "tslint --project browser/tsconfig.json --config tslint.json && tslint vim/core/oni-plugin-typescript/src/**/*.ts",
     "lint:main": "tslint --project main/tsconfig.json --config tslint.json",
     "lint:test": "tslint --project test/tsconfig.json --config tslint.json && tslint vim/core/oni-plugin-typescript/src/**/*.ts",
+    "pack": "build --publish never",
     "start": "concurrently --kill-others \"npm run start-hot\" \"npm run watch:browser\" \"npm run watch:plugins\"",
     "start-hot": "cross-env NODE_ENV=development electron lib/main/src/main.js",
     "start-not-dev": "cross-env electron main.js",
