+++ test/common/index.ts
@@ -1,1 +1,2 @@
 export * from "./Oni"
+export * from "./runInProcTest
